Some bugfixes & change account data model
Account data model has now a method to check requirement data is given.
As LINQ is using the setter methods, these methods doesn't have to show
error messages. Instead, a callable method was implemented to check the
status on save.
"ApplicationBarIconButton" does not reset the focus on click. This is
necessary in order to update the data binding of current input element.
The Extension "UpdateBindingSource "does this job now.